pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "API v1"
authors = ["Diddychriss"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "3.10.12"
django = "5.0.6"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.2"
psycopg2-binary = "2.9.9"
pytest = "8.2.0"
pytest-django = "4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
